sfs_server/requirements.txt
Django>=3.1.3,<4.0.0
channels>=3.0.2,<4.0.0
djangorestframework>=3.12.2,<4.0.0
Pillow>=8.0.1,<9.0.0
psycopg2-binary>=2.8.6,<2.9.0
pytest-django>=4.1.0,<5.0.0
